Move and fix RepoService test
